requirements.txt
asgiref==3.7.2
boto3==1.29.0
botocore==1.32.0
Django==4.2.7
django-environ==0.11.2
djangorestframework==3.14.0
jmespath==1.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
